+++ tests/ocr_tsl/test_tsl.py
@@ -68,24 +68,16 @@ def test_get_tsl_model(monkeypatch):
     monkeypatch.setattr(tsl, 'TSL_MODEL_OBJ', 'test')
     assert tsl.get_tsl_model() == 'test'
 
-def test_unload_tsl_model_if_loaded(monkeypatch):
+def test_unload_tsl_model_if_loaded(monkeypatch, mock_base_model):
     """Test unload tsl model is called if load with an already loaded model."""
-    class A(): # pylint: disable=missing-class-docstring,invalid-name
-        def __init__(self):
-            self.load_called = False
-            self.unload_called = False
-        def load(self): # pylint: disable=missing-function-docstring
-            self.load_called = True
-        def unload(self): # pylint: disable=missing-function-docstring
-            self.unload_called = True
-    a = A() # pylint: disable=invalid-name
-    b = A() # pylint: disable=invalid-name
-    monkeypatch.setattr(tsl, 'TSL_MODEL_OBJ', a)
+    base1 = mock_base_model() # pylint: disable=invalid-name
+    base2 = mock_base_model() # pylint: disable=invalid-name
+    monkeypatch.setattr(tsl, 'TSL_MODEL_OBJ', base1)
     monkeypatch.setattr(tsl, 'TSL_MODEL_ID', 'test')
-    monkeypatch.setattr(m.TSLModel, 'from_entrypoint', lambda *args, **kwargs: b)
+    monkeypatch.setattr(m.TSLModel, 'from_entrypoint', lambda *args, **kwargs: base2)
     tsl.load_tsl_model('test2')
 
-    assert not a.load_called
-    assert a.unload_called
-    assert b.load_called
-    assert not b.unload_called
+    assert not base1.load_called
+    assert base1.unload_called
+    assert base2.load_called
+    assert not base2.unload_called
